+++ lib/sorbet-coerce/configuration.rb
